requirements.txt
azure-cognitiveservices-vision-customvision
cognitive_face
Pillow==2.9.0